Adds a safety check to the WipeEnvironment task to help prevent running against the wrong environment accidentally.

+++ ContinuousIntegration/Run-Ci.ps1
@@ -240,6 +240,11 @@ task WipeEnvironment `
     -depends LoadEnvironmentConfigSettings `
     -description "Removes all data and services to bring the environment back to a clean state." {
 
+    $answer = Read-Host `
+        -Prompt ("This will permanently delete all services and data from the '{0}' environment, are you sure?" -f $AppSettings.EnvironmentName)
+
+    if ($answer -ne "y") { return }
+
     $siteName = $AppSettings.AzureWebsiteServiceName
     Write-Host "Removing Azure Website $siteName..." -NoNewline
     Remove-AzureWebsite -Name $siteName -Force -ErrorAction SilentlyContinue | Out-Null
